Fix dashboard header
